Unterstützt nun ".wav" wiedergabe
